[ANDROSDK-1962] Move api logic to network package
