+++ core/src/main/java/org/hisp/dhis/android/core/option/OptionGroupCollectionRepository.kt
@@ -33,9 +33,9 @@ import org.hisp.dhis.android.core.arch.repositories.collection.internal.ReadOnly
 import org.hisp.dhis.android.core.arch.repositories.filters.internal.FilterConnectorFactory
 import org.hisp.dhis.android.core.arch.repositories.filters.internal.StringFilterConnector
 import org.hisp.dhis.android.core.arch.repositories.scope.RepositoryScope
-import org.hisp.dhis.android.core.option.internal.OptionGroupFields
 import org.hisp.dhis.android.core.option.internal.OptionGroupOptionChildrenAppender
 import org.hisp.dhis.android.core.option.internal.OptionGroupStore
+import org.hisp.dhis.android.network.optiongroup.OptionGroupFields
 import org.koin.core.annotation.Singleton
 
 @Singleton
