+++ core/src/main/java/org/hisp/dhis/android/network/common/dto/ObjectWithUidDTO.kt
@@ -30,8 +30,13 @@ package org.hisp.dhis.android.network.common.dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
+import org.hisp.dhis.android.core.common.ObjectWithUid
 
 @Serializable
 internal data class ObjectWithUidDTO(
     @SerialName("id") val uid: String,
-)
+) {
+    fun toDomain(): ObjectWithUid {
+        return ObjectWithUid.create(uid)
+    }
+}
