+++ core/src/main/java/org/hisp/dhis/android/network/option/OptionNetworkHandlerImpl.kt
@@ -43,6 +43,6 @@ internal class OptionNetworkHandlerImpl(
     override suspend fun getOptions(optionUids: Set<String>, page: Int, pageSize: Int): PayloadJson<Option> {
         val optionSetUidsFilterStr = "optionSet." + ObjectWithUid.uid.`in`(optionUids).generateString()
         val apiPayload = service.getOptions(OptionFields.allFields, optionSetUidsFilterStr, true, page, pageSize)
-        return apiPayload.mapItems(::optionDtoToDomainMapper)
+        return apiPayload.mapItems(OptionDTO::toDomain)
     }
 }
