+++ core/src/main/java/org/hisp/dhis/android/network/optiongroup/OptionGroupService.kt
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2004-2023, University of Oslo
+ *  Copyright (c) 2004-2024, University of Oslo
  *  All rights reserved.
  *
  *  Redistribution and use in source and binary forms, with or without
@@ -25,21 +25,20 @@
  *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.android.core.option.internal
+package org.hisp.dhis.android.network.optiongroup
 
-import org.hisp.dhis.android.core.arch.api.HttpServiceClient
-import org.hisp.dhis.android.core.arch.api.payload.internal.PayloadJackson
 import org.hisp.dhis.android.core.option.OptionGroup
+import org.hisp.dhis.android.network.common.HttpServiceClientKotlinx
 import org.hisp.dhis.android.network.common.fields.Fields
 import org.koin.core.annotation.Singleton
 
 @Singleton
-internal class OptionGroupService(private val client: HttpServiceClient) {
-    suspend fun optionGroups(
+internal class OptionGroupService(private val client: HttpServiceClientKotlinx) {
+    suspend fun getOptionGroups(
         fields: Fields<OptionGroup>,
         dataSetUidsFilter: String,
         paging: Boolean,
-    ): PayloadJackson<OptionGroup> {
+    ): OptionGroupPayload {
         return client.get {
             url("optionGroups")
             parameters {
